Add cluster update support (#14)
Description of changes:
Adds hooks for checking cluster status, and gating update or delete operations unless cluster is in `ACTIVE` status. Adds custom update logic for the `UpdateClusterVersion` and `UpdateClusterConfig` operations. 

Note:
Update operations are asynchronous for EKS clusters, but can only be made one at a time. At the end of any update call, we forcefully set the synced condition to false and requeue with a 15 second duration. This gives the server time to receive and start the async update request, start the update and set the cluster status to `UPDATING` before we requeue and call read one.

By submitting this pull request, I confirm that my contribution is made under
the terms of the Apache 2.0 license.