Update `acktest` to `589f834` (#56)
Description of changes:
The current version of boto3 uses `1.22` as the latest version for EKS clusters. None of the e2e tests define a version, so they all fall back to whichever default boto3 supplies - which will be the latest version it knows about. This PR upgrades the `acktest` dependency, which now points to a much more recent version of `boto3` which supports `1.24` and ensures tests are running against the most up-to-date EKS cluster version.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.