Introduce annotation to tweak `scalingConfig.desiredSize` control (#85)
Introduce Annotation for Dynamic Nodegroup `spec.ScalingConfig.desiredSize`
Control

This patch introduces a new annotation that allows flexible management of the
desired size for a nodegroup scaling configuration. The annotation,
`eks.services.k8s.aws/desired-size-managed-by`, can be set on a nodegroup custom
resource and specifies whether the desired size is managed externally by an
autoscaler or by the ACK controller.

The possible values for the annotation include 'external-autoscaler' and
'ack-eks-controller'. If the annotation is not set or has an invalid value, the
controller defaults to managing the desired size.

Adjustments are made to delta comparison logic, custom pre and post-compare
hooks, and the update configuration logic to ensure proper handling of the
desired size based on the annotation.

This enhancement provides flexibility in managing the desired size of a
nodegroup, allowing external entities or the ACK controller to control this
aspect.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.