Add `AccessConfig` to `Cluster` CRD (#88)
This patch introduces a new field to the Cluster CRD Spec, named
`AccessConfig`. This field allows setting two new parameters in a
cluster:
- Before creation, a users will be able to set
  `bootstrapClusterCreatorAdminPermissions`
- `authenticationMode` after and after the cluster creation.

This patch also adds e2e tests to ensure the proper functioning of this
feature.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.