Allow tweaking NodeGroupVersion update requests.
Prior to this patch, the eks-controller wasn't able to set the
ReleaseVersion, LaunchTemplate or the force the attributes when making
an `UpdateNodeGroupVersion` API Call.

This patch correctly sets the ReleaseVersion and LaunchTemplate in the
API request, and introduces a new annotation allowing users to set the
Force attribute to true.

We're adding a new annotation
`eks.services.k8s.aws/force-update-version` - when set to true, the
controller will toggle the Force boolean in the update request.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>