Correctly retrieve the ARN for `PodIdentityAssociation`
In the previous patch, we added tag update support to all EKS resources.
In the process, we also began populating the PIA resource ARN in the
`ackResourcemetadata` field. The tag update logic, assumed that all
resouces would have an ARN in the `ackResourceMetadata`, which is not
the case for old resources (created by an eks-controller < 1.2.2).

This patch correctly retrieves the ARN from the appropriate location.
All old resources will have their ARN directly popualted in
`ackResourceMetadata` once the controller reconciles.

This patch also publishes `v1.2.3` release manifests

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>