Refine update patterns for `Cluster` and `Nodegroup` resources
Prior to this patch, we observed that there was some missing logic for
handling `Cluster` and `Nodegroup` updates, especially when dealing with
version upgrades - making it a painful/incomplete UX when using the
eks-controller.

This commit enhances the logic for managing updates and upgrades of
cluster and nodegroups resources, enabling users to experience a better
declarative resource model. Here's what the controller can do now:

`Cluster` resources:
- Introduces `Cluster` version rolling update, allowing users to specify
  any version higher than the "current" one, even if the EKS API doesn't
  allow it. The controller will incrementally upgrade the the cluster
  version until it reaches the desired state. e.g `1.25` -> `1.29`

`Nodegroup` resources:
- Allows updates for `ReleaseVersion` and `LaunchTemplate`
- Enhances the delta/comparison logic for both `Version` and
  `ReleaseVersion` fields.
- Carefully crafts the `UpdateNodeGroupVersion` API input to avoid
  situations where the controller/resource can get trapped in a infinite
  update loop (`1.27` -> `1.28` -> `1.27`)
- Set `DiskSize`, `RemoteAccess`, `NodeRole` and `Subnets` as immutable
  fields

Additionally this commit is also adding e2e tests for the features/bug
fixes mentionned above

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>