Enhance EKS `AccessEntries` policy attachements and delta logic
Prior to this patch, we observed some issues with the controller
behaviour regarding access policy association and delta computations,
specifically:
- False positive deltas were detected at every reconciliation (or
  controller restart)
- When a delta was detected, the controller disociate the latest
  policies, and patch the desired ones. (Even when it's unnecessary).

The root cause was identifier as the usage of `reflect.DeepEqual` for
slices. According to the `reflect` library documentation https://pkg.go.dev/reflect#DeepEqual
`DeepEqual`  returns false if the comapred elements are a nil slice and
a "non-nil empty slice" (e.g `nil` and `[]string{}`).
This combined with the possibility of a user providing a nil
`namespaces` slice, while the controller sets an empty one (based on the
API response), cause behaviour inconsistencies.

Users shouldn't care/worry about the "nil-ity" of their arrays/slices;
it is up to the controller to accurately compute the deltas.

This patch replaces the usage of `reflect.DeepEqual` with a meticulous
custom delta function.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>