Enable eks-controller to update `Cluster` encryption configuration
This commits adds the capability fo rthe eks-controller to update the
encryption configuration for an EKS Cluster. Only clusters that didn't
receive a encryption key on create are able to be updated.

If a user tries to modify or delete an existing encryption configuraiton
the controller will set a terminal condition stating that it's an
impossible operation.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>