Fix rolling updates flaky tests.
Recently we added a new feature that allowed cluster rolling updates,
allowing users to request 1.25->1.29 cluster upgrades. The tests of this
feature consisted of creating a 1.27 cluster, patching 1.29 as version
and then testing that the cluster goes from 1.27 to 1.28 then 1.29.
However we realized that sometimes when checking that the cluster is in
1.28, we find that it's actually 1.29... meaning that the waiter didn't
get the chance to find the cluster in a an active_state...

This patch reduce the waiter delay to quickly catch the cluster in a
active_state

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>