Fix the serviceAccount print column (#111)
Issue #, if available: N/A

Description of changes: Fixes the printcolumn on podidentityassociations.  The ServiceAccount field is set to:
```
ServiceAccount *string `json:"serviceAccount"`
```

But the printcolumn was set to:

```
// +kubebuilder:printcolumn:name="SERVICEACCOUNT",type=string,priority=0,JSONPath=`.spec.serviceaccount`
```

It is case sensitive so it was not matching.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.