Allow to update subnets and security group IDs (#119)
Issue #, if available:
https://github.com/aws-controllers-k8s/community/issues/1855

Description of changes:
Subnets or security group IDs can (now) be updated with the api but not at the same time as the public or private accesses.
This patch split the api calls for theses changes into two separate methods.

This also fix https://github.com/aws-controllers-k8s/community/issues/1855 as the changes in subnets or security group IDs triggers an update but these field are removed before the call to the api. The api then response with an error because there is no changes to apply.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.