Ensure Ref is not overwritten in Spec
Similarly as the ec2-controller, the ref stored in a struct
is being overwritten during create/update. With this fix
we ensure the ref would not be discarded.

Currently there isn't a way to test this fix, as we can't
run two controllers simultaneously and reference subnet names
in cluster directly.
Instead I will document an example with the fix working in
this PR