Add support for EKS Cluster `UpdatePolicy`
This commit regenerates the controller using the `v1.55.5` version of the
AWS sdk go, adding the `BootstrapSelfManagedAddons` and `UpgradePolicy`
fields to the `ClusterSpec`. It also adds hooks to handle `UpgradePolicy` updates
and an E2E test for testing the new functionality.

This patch also explicetly ignores differences in the
`BootstrapSelfManagedAddons` field since it's a create only field.