Set default cluster UpgradePolicy if nil
Currently, when a cluster has been created with an older
eks controller, and tries to reconcile with a new one,
UpgradePolicy is set to nil, and panics with nulll pointer
exception during updateClusterUpgradePolicy operation.

As a fix, I am considering setting the upgradePolicy to
the value that is retured from AWS to ensure user's
preset option is maintained.

This change will not change the customer's spec,
until the users deside to define/add the field
themselves.