chore: Set cluster status to updating after successful update
Currently, when we successfully update an EKS cluster, trigger a
controller 15 second requeue.
During these 15 seconds, the actual cluster Status does not get updated,
meaning, if previously it was Active, it will remain Active until the
end of the 15 seconds where the controller will change it to Updating.

With this change, we will be setting the Updating status ourselves when
the update api call is successful, and during the next reconciliation
(still 15 seconds) the controller will change it to whatever status is
returned by the AWS API.