Ignore Node Version and EKS Version check if AMI Type is bottlerocket (#179)
Issue #, if available:
- https://github.com/aws-controllers-k8s/community/issues/2644

Description of changes:
Since the bottlerocket versions don't align with EKS Versions, we should ignore the check of comparing nodeversion and eks version. Otherwise, it will result in error similar to `version and release version do not match: 1.32 and 1.47`

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.