chore: remove InvalidParameterException from terminal exceptions (#192)
Issue [#1844](https://github.com/aws-controllers-k8s/community/issues/1844)

Description of changes:
InvalidParameterException was returned when the controller tried
creating an EKS cluster with a roleARN that does not exist. This case
should not be considered terminal, as a role may eventually be created.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.